make tealfmt usable as a module
